reorganize test utilities
