+++ tests/tensor_base/test_pow_special_cases.py
@@ -9,7 +9,8 @@ from hypothesis import given
 import mygrad as mg
 from mygrad.math.arithmetic.ops import Positive, Square
 from tests.custom_strategies import tensors
-from tests.utils import adds_constant_arg, expected_constant
+from tests.utils.checkers import expected_constant
+from tests.utils.wrappers import adds_constant_arg
 
 from ..wrappers.uber import backprop_test_factory, fwdprop_test_factory
